Extracted the duplicated directory fixture builder into 'DirectoryFixtureTrait'.
